Enabling caching. Oops.
The server was not setting any cache directives at all, and Volley interprets
this to mean not to cache (which is a reasonable, conservative approach.) So,
this CL adds cache directives, and also puts the ID for photo lookup path
(i.e. /photos/<event-id>) so that each event has a unique URL and are thereby
cacheable at all.