Configure Android app via QR code generated by server.
Adds a server URL path handler for /qrconfig that generates a QR code packed
with a Java properties block, intended to be scanned by a phone app to
populate its configuration. Previously, this config information was hardcoded
into the .apk. This makes development (as well as) publication/distribution
harder, as it means replacing files and rebuilding a .apk to switch to a new
server (e.g. prod vs. dev, or for others if someone picks up the code.)

So, now on first launch the app will automatically launch Barcode Scanner (or
prompt to install it and then launch it), and then load its configuration by
parsing the resulting Java properties string.

It's worth noting that the QR code includes the actual base64-encoded
server (pub-key) cert keystore. Since the app supports self-signed certs,
there needs to be a way to load the BKS keystore as config data. I thought
about having the server simply serve its own BKS, and have the QR properties
block include only a URL to the BKS; but there is no way to do this without a
MITM vulnerability. Fortunately the BKS can fit into a QR code along with the
other info, at least for a small BKS with just one cert pub-key.