Update authToken verification to latest spec.
Turns out the OAuth2 "hybrid app" thing was based on an "early version"
of the OAuth2 spec, that later changed. Google updated the JWTs they
issue to use the later spec, which was causing the authToken validation
to fail. Of course, it would have been nice if the docs had told me
I was using an early version, but it's fixed now. Thanks, Obama!